Extract WebviewHtmlGenerator and AssetLoader from GitGraphView
